Addressing other PR comments
