+++ packages/catalog-process/src/services/catalogService.ts
@@ -2,13 +2,11 @@ import {
   AuthData,
   DB,
   FileManager,
-  RiskAnalysisValidatedForm,
   eventRepository,
   hasPermission,
   logger,
   riskAnalysisValidatedFormToNewRiskAnalysis,
   userRoles,
-  validateRiskAnalysis,
 } from "pagopa-interop-commons";
 import {
   Descriptor,
@@ -23,7 +21,6 @@ import {
   catalogEventToBinaryData,
   descriptorState,
   generateId,
-  operationForbidden,
   unsafeBrandId,
   ListResult,
   AttributeId,
@@ -32,7 +29,6 @@ import {
   eserviceMode,
   Tenant,
   RiskAnalysis,
-  TenantKind,
 } from "pagopa-interop-models";
 import { match } from "ts-pattern";
 import {
@@ -75,7 +71,6 @@ import {
 import { config } from "../utilities/config.js";
 import { nextDescriptorVersion } from "../utilities/versionGenerator.js";
 import {
-  draftDescriptorAlreadyExists,
   eServiceCannotBeDeleted,
   eServiceCannotBeUpdated,
   eServiceDescriptorNotFound,
@@ -88,44 +83,19 @@ import {
   attributeNotFound,
   inconsistentDailyCalls,
   originNotCompliant,
-  dailyCallsCannotBeDecreased,
-  eserviceNotInDraftState,
-  eserviceNotInReceiveMode,
-  tenantKindNotFound,
   tenantNotFound,
-  riskAnalysisValidationFailed,
 } from "../model/domain/errors.js";
 import { formatClonedEServiceDate } from "../utilities/date.js";
 import { ReadModelService } from "./readModelService.js";
-
-const assertRequesterAllowed = (
-  producerId: TenantId,
-  requesterId: TenantId
-): void => {
-  if (producerId !== requesterId) {
-    throw operationForbidden;
-  }
-};
-
-function assertIsDraftEservice(eservice: EService): void {
-  if (eservice.descriptors.some((d) => d.state !== descriptorState.draft)) {
-    throw eserviceNotInDraftState(eservice.id);
-  }
-}
-
-function assertIsReceiveEservice(eservice: EService): void {
-  if (eservice.mode !== eserviceMode.receive) {
-    throw eserviceNotInReceiveMode(eservice.id);
-  }
-}
-
-function assertTenantKindExists(
-  tenant: Tenant
-): asserts tenant is Tenant & { kind: NonNullable<Tenant["kind"]> } {
-  if (tenant.kind === undefined) {
-    throw tenantKindNotFound(tenant.id);
-  }
-}
+import {
+  assertRequesterAllowed,
+  assertDailyCallsAreConsistentAndNotDecreased,
+  assertIsDraftEservice,
+  assertIsReceiveEservice,
+  assertTenantKindExists,
+  validateRiskAnalysisOrThrow,
+  assertHasNoDraftDescriptor,
+} from "./validators.js";
 
 const retrieveEService = async (
   eserviceId: EServiceId,
@@ -244,15 +214,6 @@ const archiveDescriptor = (
   logger.info(`Archiving Descriptor ${descriptor.id} of EService ${streamId}`);
 
   return updateDescriptorState(descriptor, descriptorState.archived);
-};
-
-const hasNotDraftDescriptor = (eservice: EService): void => {
-  const hasDraftDescriptor = eservice.descriptors.some(
-    (d: Descriptor) => d.state === descriptorState.draft
-  );
-  if (hasDraftDescriptor) {
-    throw draftDescriptorAlreadyExists(eservice.id);
-  }
 };
 
 const replaceDescriptor = (
@@ -764,7 +725,7 @@ export function catalogServiceBuilder(
 
       const eservice = await retrieveEService(eserviceId, readModelService);
       assertRequesterAllowed(eservice.data.producerId, authData.organizationId);
-      hasNotDraftDescriptor(eservice.data);
+      assertHasNoDraftDescriptor(eservice.data);
 
       const newVersion = nextDescriptorVersion(eservice.data);
 
@@ -1399,40 +1360,5 @@ const applyVisibilityToEService = (
     ),
   };
 };
-
-function assertDailyCallsAreConsistentAndNotDecreased({
-  dailyCallsPerConsumer,
-  dailyCallsTotal,
-  updatedDailyCallsPerConsumer,
-  updatedDailyCallsTotal,
-}: {
-  dailyCallsPerConsumer: number;
-  dailyCallsTotal: number;
-  updatedDailyCallsPerConsumer: number;
-  updatedDailyCallsTotal: number;
-}): void {
-  if (updatedDailyCallsPerConsumer > updatedDailyCallsTotal) {
-    throw inconsistentDailyCalls();
-  }
-  if (
-    updatedDailyCallsPerConsumer < dailyCallsPerConsumer ||
-    updatedDailyCallsTotal < dailyCallsTotal
-  ) {
-    throw dailyCallsCannotBeDecreased();
-  }
-}
-
-function validateRiskAnalysisOrThrow(
-  riskAnalysisForm: EServiceRiskAnalysisSeed["riskAnalysisForm"],
-  schemaOnly: boolean,
-  tenantKind: TenantKind
-): RiskAnalysisValidatedForm {
-  const result = validateRiskAnalysis(riskAnalysisForm, schemaOnly, tenantKind);
-  if (result.type === "invalid") {
-    throw riskAnalysisValidationFailed(result.issues);
-  } else {
-    return result.value;
-  }
-}
 
 export type CatalogService = ReturnType<typeof catalogServiceBuilder>;
