+++ packages/catalog-process/src/services/validators.ts
@@ -1,0 +1,97 @@
+import {
+  RiskAnalysisValidatedForm,
+  validateRiskAnalysis,
+} from "pagopa-interop-commons";
+import {
+  TenantId,
+  operationForbidden,
+  EService,
+  descriptorState,
+  eserviceMode,
+  Tenant,
+  TenantKind,
+  Descriptor,
+} from "pagopa-interop-models";
+import {
+  eserviceNotInDraftState,
+  eserviceNotInReceiveMode,
+  tenantKindNotFound,
+  inconsistentDailyCalls,
+  dailyCallsCannotBeDecreased,
+  riskAnalysisValidationFailed,
+  draftDescriptorAlreadyExists,
+} from "../model/domain/errors.js";
+import { EServiceRiskAnalysisSeed } from "../model/domain/models.js";
+
+export function assertRequesterAllowed(
+  producerId: TenantId,
+  requesterId: TenantId
+): void {
+  if (producerId !== requesterId) {
+    throw operationForbidden;
+  }
+}
+
+export function assertIsDraftEservice(eservice: EService): void {
+  if (eservice.descriptors.some((d) => d.state !== descriptorState.draft)) {
+    throw eserviceNotInDraftState(eservice.id);
+  }
+}
+
+export function assertIsReceiveEservice(eservice: EService): void {
+  if (eservice.mode !== eserviceMode.receive) {
+    throw eserviceNotInReceiveMode(eservice.id);
+  }
+}
+
+export function assertTenantKindExists(
+  tenant: Tenant
+): asserts tenant is Tenant & { kind: NonNullable<Tenant["kind"]> } {
+  if (tenant.kind === undefined) {
+    throw tenantKindNotFound(tenant.id);
+  }
+}
+
+export function assertDailyCallsAreConsistentAndNotDecreased({
+  dailyCallsPerConsumer,
+  dailyCallsTotal,
+  updatedDailyCallsPerConsumer,
+  updatedDailyCallsTotal,
+}: {
+  dailyCallsPerConsumer: number;
+  dailyCallsTotal: number;
+  updatedDailyCallsPerConsumer: number;
+  updatedDailyCallsTotal: number;
+}): void {
+  if (updatedDailyCallsPerConsumer > updatedDailyCallsTotal) {
+    throw inconsistentDailyCalls();
+  }
+  if (
+    updatedDailyCallsPerConsumer < dailyCallsPerConsumer ||
+    updatedDailyCallsTotal < dailyCallsTotal
+  ) {
+    throw dailyCallsCannotBeDecreased();
+  }
+}
+
+export function assertHasNoDraftDescriptor(eservice: EService): void {
+  const hasDraftDescriptor = eservice.descriptors.some(
+    (d: Descriptor) => d.state === descriptorState.draft
+  );
+  if (hasDraftDescriptor) {
+    throw draftDescriptorAlreadyExists(eservice.id);
+  }
+}
+
+export function validateRiskAnalysisOrThrow(
+  riskAnalysisForm: EServiceRiskAnalysisSeed["riskAnalysisForm"],
+  schemaOnly: boolean,
+  tenantKind: TenantKind
+): RiskAnalysisValidatedForm {
+  const result = validateRiskAnalysis(riskAnalysisForm, schemaOnly, tenantKind);
+  if (result.type === "invalid") {
+    throw riskAnalysisValidationFailed(result.issues);
+  } else {
+    return result.value;
+  }
+}
